refactor(cli): organize migrated workflow modules
