+++ src/lib/domain/README.md
@@ -1,0 +1,21 @@
+<!-- SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved. -->
+<!-- SPDX-License-Identifier: Apache-2.0 -->
+
+# Domain helpers
+
+Domain modules contain pure policy and decision logic. They should not import oclif, spawn processes, read host state directly, or call Docker/OpenShell.
+
+Preferred layout:
+
+```text
+src/lib/domain/<area>/<topic>.ts
+src/lib/domain/<area>/<topic>.test.ts
+```
+
+Flat files are acceptable for small cross-cutting helpers, but workflow-specific logic should live under an area directory that matches the command/action stack when practical:
+
+```text
+src/commands/internal/<area>/<verb>.ts
+src/lib/actions/<area>/<verb>.ts
+src/lib/domain/<area>/<topic>.ts
+```
